backend/management/urls.py
from django.urls import path, include
from rest_framework.routers import DefaultRouter
from .views import StudentViewSet, TeacherViewSet, CourseViewSet, AttendanceRecordViewSet, UserViewSet, LoginView, LogoutView,ResetPasswordView
from rest_framework_simplejwt.views import TokenRefreshView

router = DefaultRouter()
router.register(r'students', StudentViewSet)
router.register(r'teachers', TeacherViewSet)
router.register(r'courses', CourseViewSet, basename='course')
router.register(r'attendance', AttendanceRecordViewSet,basename='attendance')
router.register(r'users', UserViewSet, basename='user')

urlpatterns = [
    path('', include(router.urls)),
    path('login/', LoginView.as_view(), name='login'),
    path("logout/", LogoutView.as_view()),
    path("token/refresh/", TokenRefreshView.as_view()),
    path("reset-password/", ResetPasswordView.as_view()),
    path("api/", include(router.urls)),
    ]
